+++ src/toolkit/res_buf.h
@@ -61,7 +61,7 @@ typedef std::vector<Product::Ptr> ProdVec;
 template <class T>
 class ResBuf {
  public:
-  ResBuf(bool is_bulk=false, bool unpackaged=true) : cap_(INFINITY), qty_(0), is_bulk_(is_bulk), unpackaged_(unpackaged) { }
+  ResBuf(bool is_bulk=false, bool keep_packaging=false) : cap_(INFINITY), qty_(0), is_bulk_(is_bulk), keep_packaging_(keep_packaging) { }
 
   virtual ~ResBuf() {}
 
@@ -84,6 +84,13 @@ class ResBuf {
     }
     cap_ = cap;
   }
+
+  /// Sets whether the buffer should keep packaged resources
+  void keep_packaging(bool keep_packaging) {
+    keep_packaging_ = keep_packaging;
+  }
+
+  bool keep_packaging() const { return keep_packaging_; }
 
   /// Returns the total number of constituent resource objects
   /// in the buffer. Never throws.
@@ -267,7 +274,7 @@ class ResBuf {
 
     if (!is_bulk_  || rs_.size() == 0) {
       // strip package id and set as default
-      if (unpackaged_) {
+      if (!keep_packaging_) {
         m->ChangePackage();
       }
       rs_.push_back(m);
@@ -319,7 +326,7 @@ class ResBuf {
 
     for (int i = 0; i < rss.size(); i++) {
       if (!is_bulk_ || rs_.size() == 0) {
-        if (unpackaged_) {
+        if (!keep_packaging_) {
           rss[i]->ChangePackage();
         }
         rs_.push_back(rss[i]);
@@ -359,7 +366,7 @@ class ResBuf {
   bool is_bulk_;
   /// Whether materials should be stripped of their packaging before being
   /// pushed onto the resbuf. If res_buf is bulk, this is assumed true.
-  bool unpackaged_;
+  bool keep_packaging_;
 
   /// List of constituent resource objects forming the buffer's inventory
   std::list<typename T::Ptr> rs_;