+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Since last release
 * Material sell policy can package materials (#1749, #1774)
 * Use miniforge for conda CI builds instead of miniconda (#1763)
 * Warning and limits on number of packages that can be created from a resource at once (#1771)
+* Use keep_packaging instead of unpackaged in ResBuf (#1778)
 
 **Removed:**